Refactored server.js route handlers to different files. Weather now checks if database data is recent enough to send.
