+++ lib/client.js
@@ -1,0 +1,9 @@
+'use strict';
+
+require('dotenv').config();
+
+const pg = require('pg');
+const client = new pg.Client(process.env.DATABASE_URL);
+client.on('error', err=>console.error(err));
+
+module.exports = client;
